Cleaned up and renamed the routing. I may change this in the future but not now.
